+++ integration-tester/tests/load-balancer/loadBalancer.js
@@ -5,10 +5,11 @@ const infra = infrastructure.createTestInfrastructure();
 
 const containers = [];
 for (let i = 0; i < 4; i += 1) {
-  containers.push(new kelda.Container('web', 'nginx:1.10').withFiles({
-    '/usr/share/nginx/html/index.html':
-        `I am container number ${i.toString()}\n`,
-  }));
+  containers.push(new kelda.Container('web', 'nginx:1.10', {
+    filepathToContent: {
+      '/usr/share/nginx/html/index.html':
+      `I am container number ${i.toString()}\n`,
+    } }));
 }
 containers.forEach(container => container.deploy(infra));